Move skin fetching to thunk
